allow more interfaces related to the xprv 3rd bit
normalize_bytes is gone and replaced by
normalize_bytes_ed25519() and normalize_bytes_force() depending
if the paper algorithm is followed to the letter or that the
straightforward clear the 3rd highest bit regardless is used.

from_bytes also now accept xprv with no 3rd high bit set, given
that it's a legal (although unlikely) possible xprv after a
large number of derivation.